backend/django-reports/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
firebase-admin==6.5.0
google-cloud-storage==2.13.0